Added small test cases to establish persistence caused by Selenium versus model updates
The changes made in `...model` are, as the name suggests, made directly to the model with `self.create_data_group(data_source=ds)`

The changes in `...ui` are made by the Selenium browser. It clicks the necessary buttons to add a `DataSource`.

In the `...ui` method, the changes made by the Selenium driver linger beyond the scope of the test method