Change DSSToxSubstance references to DSSToxLookup
Big changes here.

We have moved the curation status for the ExtractedChemical object out of the DSSToxSubstance model and into the RawChem record by way of a join to the DSSToxLookup table. All the downstream code that used the curation attributes (SID, true chem, true cas) had to be revised. This included several of the elasticsearch indexes.