#676 Add model, migration, and unit test.
Note: this is branched off of 654, since I needed to build this model with the extratcted_text key on the RawChem object.

The migration includes a step that adds a new GroupType for the HH records.

The seed data does not add any new data groups or data sources. I used an existing one. This should probably be changed in a subsequent commit, but I want to wait until the cleanup ticket is completed.