Cumulative product count
Add a new `cumulative_product_count` property method to the PUC object. It returns the number of products associated with the PUC *or with any of its children.* The PUC csv download now includes this new property in a column, and the bubble chart refers to it for the radius.